requirements.txt
asttokens==3.0.1
blinker==1.9.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
comm==0.2.3
debugpy==1.8.19
decorator==5.2.1
Deprecated==1.3.1
diskcache==5.6.3
exceptiongroup==1.3.1
executing==2.2.1
Flask==3.1.2
Flask-Limiter==4.1.1
gunicorn==23.0.0
idna==3.11
ipykernel==7.1.0
ipython==8.37.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.3
jupyter_client==8.7.0
jupyter_core==5.9.1
limits==5.6.0
llama_cpp_python==0.2.20
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
nest-asyncio==1.6.0
numpy==2.2.6
ordered-set==4.1.0
packaging==25.0
parso==0.8.5
platformdirs==4.5.1
prompt_toolkit==3.0.52
psutil==7.1.3
pure_eval==0.2.3
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pyzmq==27.1.0
requests==2.32.5
scikit-learn==1.7.2
scipy==1.15.3
six==1.17.0
stack-data==0.6.3
threadpoolctl==3.6.0
tornado==6.5.4
traitlets==5.14.3
typing_extensions==4.15.0
urllib3==2.6.2
wcwidth==0.2.14
Werkzeug==3.1.4
wrapt==2.0.1
